refactor(core): chat sub-component started
