+++ README.md
@@ -15,6 +15,7 @@ Symfony AI consists of several lower and higher level **components** and the res
 * **Components**
   * **[Platform](src/platform/README.md)**: A unified interface to various AI platforms like OpenAI, Anthropic, Azure, Gemini, VertexAI, and more.
   * **[Agent](src/agent/README.md)**: Framework for building AI agents that can interact with users and perform tasks.
+  * **[Chat](src/chat/README.md)**: An unified interface to send messages to agents and store long-term context.
   * **[Store](src/store/README.md)**: Data storage abstraction with indexing and retrieval for AI applications.
 * **Bundles**
   * **[AI Bundle](src/ai-bundle/README.md)**: Symfony integration for AI Platform, Store and Agent components.
