+++ src/chat/.gitattributes
@@ -1,0 +1,8 @@
+/.github export-ignore
+/tests export-ignore
+.gitattributes export-ignore
+.gitignore export-ignore
+phpstan.dist.neon export-ignore
+phpunit.xml.dist export-ignore
+CLAUDE.md export-ignore
+AGENTS.md export-ignore
