+++ src/chat/.gitignore
@@ -1,0 +1,3 @@
+composer.lock
+vendor
+.phpunit.cache
