+++ src/chat/README.md
@@ -1,0 +1,25 @@
+# Symfony AI - Chat Component
+
+The Chat component provides a bridge for building chats with agents, sits on top of the Agent component,
+allowing you to create chats and submit messages to agents.
+
+**This Component is experimental**.
+[Experimental features](https://symfony.com/doc/current/contributing/code/experimental.html)
+are not covered by Symfony's
+[Backward Compatibility Promise](https://symfony.com/doc/current/contributing/code/bc.html).
+
+## Installation
+
+```bash
+composer require symfony/ai-chat
+```
+
+**This repository is a READ-ONLY sub-tree split**. See
+https://github.com/symfony/ai to create issues or submit pull requests.
+
+## Resources
+
+- [Documentation](doc/index.rst)
+- [Report issues](https://github.com/symfony/ai/issues) and
+  [send Pull Requests](https://github.com/symfony/ai/pulls)
+  in the [main Symfony AI repository](https://github.com/symfony/ai)
