+++ src/chat/src/MessageStoreInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Agent\Chat;
+namespace Symfony\AI\Chat;
 
 use Symfony\AI\Platform\Message\MessageBag;
